requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
altair>=5.0.0
requests>=2.31.0
anthropic>=0.7.0
python-dotenv>=1.0.0
